requirements.txt
repomd
pyyaml
backoff
requests
GitPython
packaging
configargparse
